requirements.txt
langchain
langchain-openai
langchain-chroma
chromadb
fastapi
uvicorn
python-dotenv
pymupdf